# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: Nuqta Backend API
  version: 1.0.0
  description: |
    REST JSON API for Nuqta POS / ERP system.
    All responses use the standard envelope:
    - Success: `{ ok: true, data: T }`
    - Error:   `{ ok: false, error: { code, message, details?, status? } }`
  contact:
    name: Nuqta Dev Team
servers:
  - url: /api/v1
    description: API v1 base path

# ─── Security Schemes ──────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT HS256 token.
        Access token TTL: 15 min (900s).
        Refresh token TTL: 7 days (604800s).
        Header: `Authorization: Bearer <accessToken>`

  # ─── Reusable Schemas ─────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_REPLAY
            - INSUFFICIENT_STOCK
            - INVALID_STATE
            - INTERNAL_ERROR
        message:
          type: string
        details:
          description: Additional context (validation field errors, etc.)
        status:
          type: integer
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"

    # ── Pagination ──
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        totalDebt: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      additionalProperties: false
    UpdateCustomerRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        isActive: { type: boolean }

    # ── Supplier ──
...
    PostingBatch:
      type: object
      properties:
        id: { type: integer }
        periodType: { type: string, enum: [day, month, year] }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        entriesCount: { type: integer }
        totalAmount: { type: integer }
        status: { type: string, enum: [draft, posted, locked] }
        postedAt: { type: string, format: date-time, nullable: true }
        postedBy: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PostPeriodRequest:
      type: object
      required: [periodType, periodStart, periodEnd]
      properties:
        periodType: { type: string, enum: [day, month, year] }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        notes: { type: string }
      additionalProperties: false

    # ── Inventory ──
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^\\d+$" }
      description: Numeric resource ID
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    SearchParam:
      name: search
      in: query
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers (paginated)
      operationId: listCustomers
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Customer" } }
                      total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Customers]
      summary: Create customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomerRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Customer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by ID
      operationId: getCustomer
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Customers]
      summary: Update customer
      operationId: updateCustomer
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCustomerRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Customer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Customers]
      summary: Delete customer
      operationId: deleteCustomer
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── CUSTOMER LEDGER ──────────────────────────────────────────────
...
  /posting/entries/{entryId}/post:
    post:
      tags: [Posting]
      summary: Post individual journal entry
      operationId: postEntry
      parameters:
        - name: entryId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Entry posted
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posting/entries/{entryId}/unpost:
    post:
      tags: [Posting]
      summary: Unpost individual journal entry
      operationId: unpostEntry
      parameters:
        - name: entryId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Entry unposted
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posting/period:
    post:
      tags: [Posting]
      summary: Batch post entries for period
      operationId: postPeriod
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostPeriodRequest" }
      responses:
        "200":
          description: Posting batch result
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/PostingBatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /posting/batches:
    get:
      tags: [Posting]
      summary: List posting batches
      operationId: listPostingBatches
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Posting batches
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/PostingBatch" } }
                      total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /posting/batches/{batchId}/lock:
    post:
      tags: [Posting]
      summary: Lock a posting batch
      operationId: lockPostingBatch
      parameters:
        - name: batchId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Batch locked
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posting/batches/{batchId}/unlock:
    post:
      tags: [Posting]
      summary: Unlock a posting batch
      operationId: unlockPostingBatch
      parameters:
        - name: batchId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Batch unlocked
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── AUDIT ────────────────────────────────────────────────────────
...